Fix 403 Forbidden on ticket reply by using loose comparison and casting for user_id in TicketController

+++ api/app/Http/Controllers/TicketController.php
@@ -31,7 +31,7 @@ class TicketController extends Controller
     {
         $ticket = Ticket::with(['user', 'messages.user'])->findOrFail($id);
 
-        if (!auth()->user()->isAdmin() && $ticket->user_id !== auth()->id()) {
+        if (!auth()->user()->isAdmin() && (int)$ticket->user_id !== (int)auth()->id()) {
             return response()->json(['message' => 'Forbidden.'], 403);
         }
 
@@ -73,7 +73,7 @@ class TicketController extends Controller
 
         $ticket = Ticket::findOrFail($id);
 
-        if (!auth()->user()->isAdmin() && $ticket->user_id !== auth()->id()) {
+        if (!auth()->user()->isAdmin() && (int)$ticket->user_id !== (int)auth()->id()) {
             return response()->json(['message' => 'Forbidden.'], 403);
         }
